refactor(credentials-mobile): split into sibling components for 500-line cap
`frontend/src/pages/Credentials/Mobile.tsx` was 812 lines, over the
500-line soft cap. Decomposed into:

- mobileTypes.ts             — types, constants, base64url helper
- CredentialCard.tsx         — single credential row card + StatusBadge
- CredentialActionSheet.tsx  — actions + set/clear-key form sheet
- LockScreen.tsx             — locked / unlocking / error UI

Main `Mobile.tsx` (now 386 lines) keeps lock-state, WebAuthn unlock,
inactivity timer, data fetching, and composes the sub-components.
Public export `CredentialsMobile` is preserved.

No behavior change.
